requirements.txt
crewai
crewai[tools]
python-dotenv
groq
requests
langchain-openai